#6 Reorganizing some stuff
